requirements.txt
playwright==1.47.0
beautifulsoup4==4.12.3
lxml==5.3.0
python-dotenv==1.0.1
